cleaned up all the warnings of mariadb, made nginx depend on wordpress and removed it entrypoint script, made the ssl certificate be made in the image creation, rather that the container, made sure to avoid creating a user in wordpress unless it is the first time starting the container

+++ srcs/requirements/wordpress/startup.sh
@@ -6,13 +6,16 @@ PHP_BIN="/usr/bin/php84"
 . /run/secrets/wp_user_pass
 . /run/secrets/wp_admin_pass
 
-export DB_PASSWORD=$DB_PASSWORD
-
-envsubst < $WP_ROOT/wp-config-sample.php > $WP_ROOT/wp-config.php
-
-export DB_PASSWORD=""
-
-sed -i "s/ = 'wp_';/\$table_prefix = 'wp_';/" $WP_ROOT/wp-config.php
+echo "initilizing wp-config"
+if [ ! -f "$WP_ROOT/wp-config.php" ]; then
+  export DB_PASSWORD=$DB_PASSWORD
+  envsubst < $WP_ROOT/wp-config-sample.php > $WP_ROOT/wp-config.php
+  sed -i "s/ = 'wp_';/\$table_prefix = 'wp_';/" $WP_ROOT/wp-config.php
+  export DB_PASSWORD=""
+  echo "wp-config was initilized"
+else
+  echo "wp-config already intilized"
+fi
 
 host=$DB_NAME
 port=$DB_PORT
@@ -28,21 +31,21 @@ until nc -z "$DB_NAME" "$port" >/dev/null 2>&1; do
     echo "Error: Database not available after $timeout seconds."
     exit 1
   fi
-  echo "Database host '$host' not yet available on port '$port'. Waiting 1 second..."
-  sleep 1
+  echo "Database host '$host' not yet available on port '$port'. Waiting 2 second..."
+  sleep 2
 done
 echo "Database is up and running!"
 
 
-
-wp core install --path=$WP_ROOT --url=$WP_SITEURL --title=inception --admin_user=$WP_ADMIN_USER --admin_email=$WP_ADMIN_EMAIL --admin_password=$WP_ADMIN_PASS
-
-wp user create $WP_USER $INTRA@student.42amman.com  --path=$WP_ROOT --role=author --user_pass=$WP_USER_PASS
-
-wp plugin install redis-cache --path=$WP_ROOT
-
-wp plugin activate redis-cache --path=$WP_ROOT
-
-wp redis enable --path=$WP_ROOT
+echo "intstalling wordpress"
+if ! wp core is-installed --path=$WP_ROOT; then
+  wp core install --path=$WP_ROOT --url=$WP_SITEURL --title=inception --admin_user=$WP_ADMIN_USER --admin_email=$WP_ADMIN_EMAIL --admin_password=$WP_ADMIN_PASS
+  wp user create $WP_USER $INTRA@student.42amman.com  --path=$WP_ROOT --role=author --user_pass=$WP_USER_PASS
+  wp plugin install redis-cache --path=$WP_ROOT
+  wp plugin activate redis-cache --path=$WP_ROOT
+  wp redis enable --path=$WP_ROOT
+else
+  echo "wordpress already installed"
+fi
 
 exec "$@"
